code-analysis/mini-ray/requirements.txt
# Mini-Ray 依赖包

# gRPC 相关
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
